+++ HISTORY.md
@@ -9,6 +9,7 @@ v1.6.3 ()
 * Fixed application serial number not updated when rename instance name
 * Echo the data compress rate every 60 seconds
 * Improve the period buffer sleep length
+* Show the error information when failed to parse app's conf.json file
 
 
 v1.6.2 (2019-09-17)